File rename/split due to class refactoring.
